unixgram transport: build for all Unix(-like) platforms again - can be used by libkrun on Linux
When using gvproxy in the vfkit unixgram mode, it can talk to real Apple Virtualization framework VMs but also libkrun which reimplements the same protocol on both macOS and Linux. For libkrun virtio-net tests, it is useful to have a cross-platform network helper with the same configuration. So gvproxy with vfkit can be that if unixgram is built for all Unix-like platforms.
