+++ pkg/transport/unixgram_windows.go
@@ -1,4 +1,4 @@
-//go:build !darwin
+//go:build windows
 
 package transport
